delete messages

+++ main.py
@@ -6,6 +6,7 @@ from commands import *
 
 token = ''
 bot = telebot.TeleBot(token)
+messages = dict()
 
 
 def creat_markup():
@@ -21,53 +22,66 @@ def creat_markup():
     return markup
 
 
-def delete_messages(*messages):
-    for msg in messages:
-        bot.delete_message(msg.chat.id, msg.message_id)
+def delete_messages(msg, msgs):
+    # delete user message
+    bot.delete_message(msg.chat.id, msg.message_id)
+    # delete previous bot messages
+    if msg.chat.id in msgs.keys():
+        for msg in msgs[msg.chat.id]:
+            try:
+                bot.delete_message(msg.chat.id, msg.message_id)
+            except:
+                print("delete error")
+        msgs[msg.chat.id] = []
 
 
 @bot.message_handler(commands=['start', 'help'])
-def start_message(message):
-    bot.send_message(message.chat.id, const.list_com + const.default_com)  # + get_commands().join("\n"))
-    bot.send_message(message.chat.id, const.choose, reply_markup=creat_markup())
+def info_message(message):
+    bot_message = bot.send_message(message.chat.id, const.list_com + const.default_com + const.choose,
+                                   reply_markup=creat_markup())  # + get_commands().join("\n"))
+    delete_messages(message, messages)
+    if bot_message.chat.id in messages.keys():
+        messages[bot_message.chat.id].append(bot_message)
+    else:
+        messages[bot_message.chat.id] = [bot_message]
 
 
 @bot.message_handler(commands=['playpause'])
 def play_pause(message):
     do_action("playpause")
-    bot.send_message(message.chat.id, const.choose, reply_markup=creat_markup())
+    info_message(message)
 
 
 @bot.message_handler(commands=['next'])
 def play_next(message):
     do_action("nexttrack")
-    bot.send_message(message.chat.id, const.choose, reply_markup=creat_markup())
+    info_message(message)
 
 
 @bot.message_handler(commands=['previous'])
 def play_previous(message):
     do_action("prevtrack")
-    bot.send_message(message.chat.id, const.choose, reply_markup=creat_markup())
+    info_message(message)
 
 
 @bot.message_handler(commands=['volumeup'])
 def volume_up(message):
     for i in range(5):
         do_action("volumeup")
-    bot.send_message(message.chat.id, const.choose, reply_markup=creat_markup())
+    info_message(message)
 
 
 @bot.message_handler(commands=['volumedown'])
 def volume_down(message):
     for i in range(5):
         do_action("volumedown")
-    bot.send_message(message.chat.id, const.choose, reply_markup=creat_markup())
+    info_message(message)
 
 
 @bot.message_handler(commands=['volumemute'])
 def volume_mute(message):
     do_action("volumemute")
-    bot.send_message(message.chat.id, const.choose, reply_markup=creat_markup())
+    info_message(message)
 
 
 @bot.message_handler(content_types=['text'])
@@ -84,5 +98,6 @@ def analize_message(message):
         volume_down(message)
     elif message.text == "🔇":
         volume_mute(message)
+
 
 bot.infinity_polling()
